requirements.txt
discord.py
openai
pytz
